+++ src/index.default.ts
@@ -14,6 +14,7 @@ import dash2Camel from './dash2Camel'
 import upperFirst from './upperFirst'
 import randomNumber from './randomNumber'
 import randomString from './randomString'
+import fingerprint from './fingerprint'
 import getCHSLength from './getCHSLength'
 import cutCHSString from './cutCHSString'
 
@@ -145,6 +146,7 @@ export default {
 	getProperty,
 	randomNumber,
 	randomString,
+	fingerprint,
 	getScrollPosition,
 	getType,
 	parseUrlParam,